+++ projects/ColorPane/src/PluginModules/PluginEnums.lua
@@ -56,10 +56,11 @@ return {
         Default = "Default",
     },
 
-    PromptError = {
+    PromptRejection = {
         InvalidPromptOptions = "InvalidPromptOptions",
         PromptAlreadyOpen = "PromptAlreadyOpen",
         ReservationProblem = "ReservationProblem",
-        PromptCancelled = "PromptCancelled"
+        PromptCancelled = "PromptCancelled",
+        SameAsInitial = "SameAsInitial",
     }
 }